Bugfix/docker minor improves

+++ Script.ps1
@@ -1,9 +1,49 @@
-docker run -it --rm -v "$(pwd):/usr/src/project" -w /usr/src/project maven:alpine mvn clean install
-Copy-Item -Path ./target/padelversus-0.0.1-SNAPSHOT.jar -Destination ./Docker/PadelVersus.jar -force
-Copy-Item -Path ./DemoImages -Destination ./Docker/DemoImages -Recurse
-cd Docker
-docker rmi padelversus
-docker image build -t padelversus -f .Dockerfile .
-docker-compose up -d
-cd ..
+# Line to run from standard cmd like intellij idea one (only windows)
+# Execution without log
+# C:/Windows/System32/WindowsPowerShell/v1.0/powershell.exe -File C:/Users/jllav/IdeaProjects/PadelVersus/Script.ps1
+# Execution with log
+# C:/Windows/System32/WindowsPowerShell/v1.0/powershell.exe -File C:/Users/jllav/IdeaProjects/PadelVersus/Script.ps1 $false
 
+# Clean and packge using local .m2 repository to do not download already gotten libraries
+docker run -it --rm -v "$(pwd):/usr/src/project" `
+                    -v"$HOME/.m2:/root/.m2" `
+                    -w /usr/src/project maven:alpine mvn clean package
+
+# If there is any problem in compilation we move and rename the .jar
+if(Test-Path ./target/padelversus-0.0.1-SNAPSHOT.jar -PathType Leaf){
+    Copy-Item -Path ./target/padelversus-0.0.1-SNAPSHOT.jar -Destination ./Docker/PadelVersus.jar -force
+}else{
+    echo "Maven compilation fail"
+    exit
+}
+
+# If there is not already moved we copy recursive the folder with images needed for the demo
+if (!(Test-Path ./Docker/DemoImages -PathType Any)){
+    Copy-Item -Path ./DemoImages -Destination ./Docker/DemoImages -Recurse
+}else{
+    Remove-Item -Recurse -Force ./Docker/DemoImages
+    Copy-Item -Path ./DemoImages -Destination ./Docker/DemoImages -Recurse
+}
+
+# We move into the folder were the .Dockerfile and docker-compose.yml is
+cd Docker
+
+# We remove the image in case we have done changes in .Dockerfile
+docker rmi padelversus
+
+# We build the app image
+docker image build -t padelversus -f .Dockerfile .
+
+# We use docker-compose to create, connect and set healthcheck containers
+if(!($args[0])){
+    docker-compose up -d
+}else{
+    docker-compose up
+}
+
+# We wait for a key pressed to stop and remove containers with docker-compose down
+read-host  "Press enter to run docker-compose down and stop execution"
+docker-compose down
+
+# We reutrned to the started directory
+cd ..
